+++ src/noahtheduke/splint/printer.clj
@@ -18,18 +18,15 @@
   (when alt
     (println "Consider using:")
     (pp/pprint alt))
-  (newline)
-  (flush))
+  (newline))
 
 (defmethod print-find "simple" [_ {:keys [filename rule-name line column message]}]
   (printf "%s:%s:%s [%s] - %s" filename line column rule-name message)
-  (newline)
-  (flush))
+  (newline))
 
 (defmethod print-find "clj-kondo" [_ {:keys [filename line column message]}]
   (printf "%s:%s:%s: warning: %s" filename line column message)
-  (newline)
-  (flush))
+  (newline))
 
 (defmethod print-find "markdown" [_ {:keys [filename rule-name form line column message alt]}]
   (println "----")
@@ -49,34 +46,41 @@
     (println "```clojure")
     (pp/pprint alt)
     (println "```")
-    (newline))
-  (flush))
+    (newline)))
 
 (defmethod print-find "json" [_ diagnostic]
   (let [diagnostic (-> diagnostic
                        (update :rule-name pr-str)
                        (update :form pr-str)
-                       (update :alt pr-str))]
+                       (update :alt pr-str)
+                       (->> (into (sorted-map))))]
     (json/write diagnostic *out* {:escape-slash false})
-    (newline)
-    (flush)))
+    (newline)))
 
 (defmethod print-find "json-pretty" [_ diagnostic]
   (let [diagnostic (-> diagnostic
                        (update :rule-name pr-str)
                        (update :form pr-str)
-                       (update :alt pr-str))]
+                       (update :alt pr-str)
+                       (->> (into (sorted-map))))]
     (json/pprint diagnostic {:escape-slash false})
-    (newline)
-    (flush)))
+    (newline)))
+
+(defmethod print-find "edn" [_ diagnostic]
+  (prn (into (sorted-map) diagnostic)))
+
+(defmethod print-find "edn-pretty" [_ diagnostic]
+  (pp/pprint (into (sorted-map) diagnostic)))
 
 (defn print-results
   [options diagnostics total-time]
   (when-not (or (:quiet options) (:silent options))
     (let [printer (get-method print-find (:output options))]
       (doseq [diagnostic (sort-by :filename diagnostics)]
-        (printer nil diagnostic))))
-  (when-not (or (:silent options) (#{"markdown" "json" "json-pretty"} (:output options)))
+        (printer nil diagnostic))
+      (flush)))
+  (when-not (or (:silent options) (#{"markdown" "json" "json-pretty"
+                                     "edn" "edn-pretty"} (:output options)))
     (printf "Linting took %sms, %s style warnings%s\n"
             total-time
             (count (remove #(= 'splint/error (:rule-name %)) diagnostics))